fix(piping): Fix Certbot not running in a terminal
Without this, Certbot would complain:

```log
Skipped user interaction because Certbot doesn't appear to be running in a terminal. You should probably include --non-interactive or --force-interactive on the command line.
```

This fix avoids needing to add this everywhere we might require user input,
as it might not be explicit and we should be using `edo` anyway.